Move schema validation to separate file
